fix review requirements
