+++ src/App.js
@@ -1,11 +1,14 @@
+import React from 'react';
 import Calculator from './components/calculator';
 
-function App() {
-  return (
-    <div className="App">
-      <Calculator />
-    </div>
-  );
-}
+class App extends React.Component {
+  render() { 
+    return (
+     <div className="App">
+       <Calculator />
+     </div>
+    );
+  }
+ }
 
 export default App;
